Copy full package out of `node_modules` (fixes #100) So now the original source is also included, not just the `dist/esm` directory
