+++ src/libs/simplex-noise/package.json
@@ -1,1 +1,81 @@
-{"type": "module"}
+{
+  "name": "simplex-noise",
+  "version": "3.0.1",
+  "description": "simplex-noise is a fast simplex noise implementation in Javascript. Works in node and in the browser.",
+  "homepage": "https://github.com/jwagner/simplex-noise.js",
+  "author": "Jonas Wagner <jonas@29a.ch> (http://29a.ch/)",
+  "main": "dist/cjs/commonjs-wrapper.js",
+  "module": "dist/esm/simplex-noise.js",
+  "exports": {
+    "require": "./dist/cjs/commonjs-wrapper.js",
+    "import": "./dist/esm/simplex-noise.js"
+  },
+  "sideEffects": false,
+  "files": [
+    "dist/esm/package.json",
+    "dist/esm/simplex-noise.js",
+    "dist/esm/simplex-noise.js.map",
+    "dist/esm/simplex-noise.d.ts",
+    "dist/cjs/simplex-noise.js",
+    "dist/cjs/simplex-noise.js.map",
+    "dist/cjs/simplex-noise.d.ts",
+    "dist/cjs/commonjs-wrapper.js",
+    "dist/cjs/commonjs-wrapper.js.map",
+    "dist/cjs/commonjs-wrapper.d.ts",
+    "simplex-noise.ts"
+  ],
+  "devDependencies": {
+    "@parcel/packager-ts": "^2.0.0-rc.0",
+    "@parcel/transformer-typescript-types": "^2.0.0-rc.0",
+    "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.7.1",
+    "@types/pngjs": "^6.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
+    "alea": "^1.0.0",
+    "benchmark": "^2.1.4",
+    "chai": "^4.3.4",
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.24.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "html-webpack-plugin": "^5.3.2",
+    "mocha": "^9.1.0",
+    "parcel": "^2.0.0-rc.0",
+    "pngjs": "^6.0.0",
+    "source-map-loader": "^3.0.1",
+    "ts-node": "^10.4.0",
+    "typedoc": "^0.22.10",
+    "typescript": "^4.5.4",
+    "webpack": "^5.51.1",
+    "webpack-cli": "^4.8.0"
+  },
+  "bugs": {
+    "url": "https://github.com/jwagner/simplex-noise.js/issues"
+  },
+  "keywords": [
+    "noise",
+    "random",
+    "simplex",
+    "plasma",
+    "procedural",
+    "generation",
+    "gfx",
+    "generative"
+  ],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jwagner/simplex-noise.js.git"
+  },
+  "scripts": {
+    "start": "parcel test/visual.html",
+    "test": "eslint simplex-noise.ts && mocha && ./test/module-compatibility.sh",
+    "build": "./build.sh",
+    "docs": "typedoc --excludePrivate --out public/docs simplex-noise.ts",
+    "prepare": "npm run-script build",
+    "benchmark": "parcel build && node ./perf/benchmark.js"
+  }
+}
